Move dialog to shared components
