+++ src/pages/Products/Inbox/index.tsx
@@ -3,7 +3,7 @@ import { Link } from '@solidjs/router';
 
 import TrayIcon from '../../../assets/icons/tray.svg';
 
-import DialogFacade from '../DialogContent/DialogFacade';
+import DialogFacade from '../../../components/DialogContent/DialogFacade';
 
 const Inbox: Component = () => {
   return (
